+++ src/talos/liftover/lift_2_0_0_to_2_1_0.py
@@ -17,6 +17,7 @@ def resultdata(data_dict: dict) -> dict:
         for variant in result['variants']:
             variant['reasons'] = variant['reasons'].pop()
     _categories = data_dict['metadata'].pop('categories')
+    data_dict['metadata']['variant_breakdown'] = {}
     data_dict['version'] = '2.1.0'
     return data_dict