+++ src/talos/utils.py
@@ -6,12 +6,13 @@ https://tenacity.readthedocs.io/en/latest/
 """
 
 import json
+import pathlib
 import re
 import string
 import zoneinfo
 from collections import defaultdict
 from datetime import datetime
-from itertools import chain, combinations_with_replacement, islice
+from itertools import chain, combinations_with_replacement, islice, combinations
 from random import choices
 from typing import TYPE_CHECKING, Any
 
@@ -72,12 +73,13 @@ DOI_URL = 'https://doi.org/'
 # don't fail, just suggest that someone raises an issue on github, but only print this message once
 PHASE_BROKEN: bool = False
 
+# a constant for the variant counting stats
+MEAN_SLASH_SAMPLE = 'Mean/sample'
+
 
 def parse_mane_json_to_dict(mane_json: str) -> dict:
     """
     Read the MANE JSON and filter it to the relevant fields
-    Args:
-        mane_json ():
 
     Returns:
         a dictionary of {Symbol: ID}
@@ -91,8 +93,6 @@ def parse_mane_json_to_dict(mane_json: str) -> dict:
 def get_random_string(length: int = 6) -> str:
     """
     get a random string of a pre-determined leng`th
-    Args:
-        length ():
 
     Returns:
         A random string comprised of upper-case letters and numbers
@@ -670,15 +670,8 @@ def read_json_from_path(read_path: str | None = None, default: Any = None, retur
 
 def get_non_ref_samples(variant: 'cyvcf2.Variant', samples: list[str]) -> tuple[set[str], set[str]]:
     """
-    for this variant, find all samples with a call
-    cyvcf2 uses 0,1,2,3==HOM_REF, HET, UNKNOWN, HOM_ALT
-
-    Args:
-        variant (cyvcf2.Variant):
-        samples (list[str]):
-
-    Returns:
-        2 sets of strings; het and hom
+    For this variant, find all samples with a call. cyvcf2 uses 0,1,2,3==HOM_REF, HET, UNKNOWN, HOM_ALT.
+    Returns 2 sets of strings; het sample ID, hom sample IDs
     """
     het_samples = set()
     hom_samples = set()
@@ -696,12 +689,7 @@ def get_non_ref_samples(variant: 'cyvcf2.Variant', samples: list[str]) -> tuple[
 
 
 def extract_csq(csq_contents: str) -> list[dict]:
-    """
-    handle extraction of the CSQ entries based on string in config
-
-    Args:
-        csq_contents ():
-    """
+    """Handle extraction of the CSQ entries based on string in config."""
 
     # allow for no CSQ data, i.e. splice variant
     if not csq_contents:
@@ -723,17 +711,14 @@ def extract_csq(csq_contents: str) -> list[dict]:
 
 def find_comp_hets(var_list: list[VARIANT_MODELS], pedigree: PedigreeParser) -> CompHetDict:
     """
-    manual implementation to find compound hets
-    variants provided in the format
+    Find compound het pairs, variants provided in the format [var1, var2, ...]
 
-    [var1, var2, ..]
-
-    generate pair content in the form
-    {sample: {var_as_string: [partner_variant, ...]}}
-
-    Args:
-        var_list (list[VARIANT_MODELS]): all variants in this gene
-        pedigree (PedigreeParser): Pedigree represetation model
+    generates pair content in the form
+    {
+        sample: {
+            var_as_string: [partner_variant, ...],
+        }
+    }
     """
 
     # create an empty dictionary
@@ -769,8 +754,7 @@ def find_comp_hets(var_list: list[VARIANT_MODELS], pedigree: PedigreeParser) -> 
 
 def generate_fresh_latest_results(current_results: ResultData):
     """
-    This will be called if a cohort has no latest results, but has
-    indicated a place to save them.
+    This will be called if a cohort has no latest results, but has indicated a place to save them.
     Args:
         current_results (ResultData): results from this current run
     """
@@ -799,9 +783,6 @@ def phenotype_label_history(results: ResultData):
     """
     Annotation in-place of the results object
     Either pull the 'date of phenotype match' from the historic data, or add 'today' to the historic data
-
-    Args:
-        results (ResultData):
     """
     # are there any history results?
     if (historic_folder := config_retrieve('result_history', None)) is None:
@@ -850,19 +831,20 @@ def phenotype_label_history(results: ResultData):
 
 
 def save_new_historic(results: HistoricVariants):
-    """
-    save the new results in the historic results dir
-
-    Args:
-        results (HistoricVariants): object to save as JSON
-    """
+    """Save the new results in the historic results dir."""
 
     if (directory := config_retrieve('result_history', None)) is None:
         logger.info('No historic data folder, no saving')
         return
 
     # we're using cloud paths here
-    new_file = to_anypath(directory) / f'{TODAY}.json'
+    dir_as_path = to_anypath(directory)
+    new_file = dir_as_path / f'{TODAY}.json'
+
+    # need a check for folder existence, or create
+    if isinstance(dir_as_path, pathlib.Path) and not dir_as_path.exists():
+        dir_as_path.mkdir(parents=True, exist_ok=True)
+
     with new_file.open('w') as handle:
         handle.write(results.model_dump_json(indent=4))
 
@@ -906,19 +888,14 @@ def annotate_variant_dates_using_prior_results(results: ResultData):
         generate_fresh_latest_results(current_results=results)
 
 
-def date_from_string(string: str) -> str:
+def date_from_string(filename: str) -> str:
     """
-    takes a string, finds the date. Simples
-    Args:
-        string (a filename):
-
-    Returns:
-        the String YYYY-MM-DD
+    Takes a filename, finds a date. Internally consistent with the way this codebase writes datetimes into file paths.
     """
-    date_search = re.search(DATE_RE, string)
+    date_search = re.search(DATE_RE, filename)
     if date_search:
         return date_search.group()
-    raise ValueError(f'No date found in {string}')
+    raise ValueError(f'No date found in {filename}')
 
 
 def find_latest_file(results_folder: str, ext: str = 'json') -> str | None:
@@ -1019,3 +996,52 @@ def date_annotate_results(current: ResultData, historic: HistoricVariants):
                     first_tagged=get_granular_date(),
                     clinvar_stars=clinvar_stars,
                 )
+
+
+def generate_summary_stats(result_set: ResultData):
+    """Reads over the variants present in the result set, and generates some per-sample and per-cohort stats."""
+
+    # make this naive, i.e. not configured specifically based on a list of Categories in configuration
+    category_count: dict[str, list[int]] = defaultdict(list)
+
+    for sample_id, sample_results in result_set.results.items():
+        if len(sample_results.variants) == 0:
+            continue
+
+        sample_variants: dict[str, set[str]] = defaultdict(set)
+
+        # iterate over all identified variants
+        for each_var in sample_results.variants:
+            var_string = each_var.var_data.coordinates.string_format
+
+            # catch all comp-het pairs as a single variant - we are electing to count comp-het events as a single
+            # variant, so we have to adjust our counting for this
+            # do this by identifying all sorted pairwise combinations, sorting them, creating a single String for each,
+            # and counting each unique String once
+            if sups := each_var.support_vars:
+                var_strings = [' '.join(sorted(combo)) for combo in combinations([var_string, *sups], 2)]
+
+            # if the variant is dominant or Hom, count it once - list of length 1
+            else:
+                var_strings = [var_string]
+
+            for unique_var in var_strings:
+                # populate the 'any's
+                sample_variants['any'].add(unique_var)
+
+                # find all categories associated with this variant. For each category, add to corresponding list and set
+                for category_value in each_var.categories:
+                    sample_variants[category_value].add(unique_var)
+
+        # record that these were the variants seen for this sample,
+        for key, set_of_strings in sample_variants.items():
+            category_count[key].append(len(set_of_strings))
+
+    # update the counts-per-sample dictionary
+    number_of_samples = len(result_set.results)
+
+    for count_list in category_count.values():
+        # pad the observed counts to the number of samples under consideration
+        count_list.extend([0] * (number_of_samples - len(count_list)))
+
+    result_set.metadata.variant_breakdown = category_count