+++ CHANGELOG.md
@@ -14,6 +14,13 @@ Suggested headings per release (as appropriate) are:
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+[7.6.0] - 2025-08-19
+
+### Changed 
+
+* No more phenopackets, instead we're using a single TSV file to contain the pedigree data, and optionally HPO terms. See the docs/Pedigree.md file for details on the new format and corresponding parser.
+* The Report now presents the variant Categories as a short phrase (e.g. "ClinVarP/LP, HighImpact", instead of "1, 3") 
+
 [7.5.0] - 2025-08-07
 
 ### Changed